requirements.txt
keras==2.15.0
tensorflow==2.15.0
numpy==1.25.2
Pillow==9.4.0